third_party/blink/web_tests/platform/generic/http/tests/devtools/profiler/heap-snapshot-summary-sorting-expected.txt

Tests sorting in Summary view of detailed heap snapshots.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testSorting

Profiler was disabled.
